Reorganize some codes for reuse among different devices.
